Adding a new implementation of the same interface, hopefully more efficient
This patch adds a new `DoubleLockBoundedQueue` class, which instead of using
just one lock for both adds and deletes, uses one lock for each, and tries to
minimize the number of operations that end up locking both.

It's worth noting that the first attempt at implementing this used two
`volatile` integers counting respectively the number of adds and the number of
deletes, the current size then being the difference between the two. Tests very
quickly made it clear that such an approach can't work, in particular when it
comes to deciding whether one operation should lock the other operation's lock
to notify sleeping threads: there can be race conditions such that consumers
and producers will effectively deadlock each other...

Also worth noting: since we do want to be able to be `Interrupt`able and want
to honor the timeout, we've replaced all the `ReentrantLock.lock()` calls with
`ReentrantLock.lockInterruptibly()`s - but we should also make the time waited
to acquire the lock count towards the timeout.

Renamed the previous single-lock implentation to `SingleLockBoundedQueue`, and
DRY-ed up the common logic between the 2 classes in `AbstractBoundedQueue`.